+++ roxbot/checks.py
@@ -25,7 +25,9 @@ SOFTWARE.
 """
 
 
+import discord
 from discord.ext import commands
+
 import roxbot
 from roxbot import guild_settings as gs
 
@@ -35,6 +37,8 @@ def is_owner_or_admin():
 	def predicate(ctx):
 		if ctx.author.id == roxbot.owner:
 			return True
+		elif isinstance(ctx.channel, discord.DMChannel):
+			return False
 		else:
 			for role in ctx.author.roles:
 				if role.id in gs.get(ctx.guild).perm_roles["admin"]:
@@ -46,6 +50,8 @@ def is_owner_or_admin():
 def _is_admin_or_mod(ctx):
 	if ctx.message.author.id == roxbot.owner:
 		return True
+	elif isinstance(ctx.channel, discord.DMChannel):
+		return False
 	else:
 		admin_roles = gs.get(ctx.guild).perm_roles["admin"]
 		mod_roles = gs.get(ctx.guild).perm_roles["mod"]
@@ -60,6 +66,8 @@ def is_admin_or_mod():
 
 
 def nsfw_predicate(ctx):
+	if isinstance(ctx.channel, discord.DMChannel):
+		return False
 	nsfw = gs.get(ctx.guild).nsfw
 	if not nsfw["channels"] and nsfw["enabled"]:
 		return nsfw["enabled"] == 1
@@ -74,4 +82,12 @@ def is_nfsw_enabled():
 
 
 def isnt_anal():
-	return commands.check(lambda ctx: gs.get(ctx.guild).is_anal["y/n"] and nsfw_predicate(ctx) or not gs.get(ctx.guild).is_anal["y/n"])
+	def predicate(ctx):
+		if isinstance(ctx.channel, discord.DMChannel):
+			return False
+		anal = gs.get(ctx.guild).is_anal["y/n"]
+		if not anal or (nsfw_predicate(ctx) and gs.get(ctx.guild).is_anal["y/n"]):
+			return True
+		else:
+			return False
+	return commands.check(predicate)